==> 4_Find_Missing_Data.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime

FOI_DEV_LIST = ['thru1997']
for i in range(1998, 2009):
    FOI_DEV_LIST.append(str(i))
FOI_DEV_LIST_2 = ['', 'Add', 'Change']
for i in range(2009, 2017):
    FOI_DEV_LIST_2.append(str(i))

baseline_col = ['BASELINE_BRAND_NAME', 'BASELINE_GENERIC_NAME', 'BASELINE_MODEL_NO', 'BASELINE_CATALOG_NO', 'BASELINE_OTHER_ID_NO', 'BASELINE_DEVICE_FAMILY', 'BASELINE_SHELF_LIFE_CONTAINED', 'BASELINE_SHELF_LIFE_IN_MONTHS', 'BASELINE_PMA_FLAG', 'BASELINE_PMA_NO', 'BASELINE_510_K__FLAG', 'BASELINE_510_K__NO', 'BASELINE_PREAMENDMENT', 'BASELINE_TRANSITIONAL', 'BASELINE_510_K__EXEMPT_FLAG', 'BASELINE_DATE_FIRST_MARKETED', 'BASELINE_DATE_CEASED_MARKETING']    
main_col = ['MDR_REPORT_KEY', 'DEVICE_EVENT_KEY', 'IMPLANT_FLAG', 'DATE_REMOVED_FLAG', 'DEVICE_SEQUENCE_NO', 'DATE_RECEIVED', 'BRAND_NAME', 'GENERIC_NAME', 'MANUFACTURER_D_NAME', 'MANUFACTURER_D_ADDRESS_1', 'MANUFACTURER_D_ADDRESS_2', 'MANUFACTURER_D_CITY', 'MANUFACTURER_D_STATE_CODE', 'MANUFACTURER_D_ZIP_CODE', 'MANUFACTURER_D_ZIP_CODE_EXT', 'MANUFACTURER_D_COUNTRY_CODE', 'MANUFACTURER_D_POSTAL_CODE', 'EXPIRATION_DATE_OF_DEVICE', 'MODEL_NUMBER', 'CATALOG_NUMBER', 'LOT_NUMBER', 'OTHER_ID_NUMBER', 'DEVICE_OPERATOR', 'DEVICE_AVAILABILITY', 'DATE_RETURNED_TO_MANUFACTURER', 'DEVICE_REPORT_PRODUCT_CODE', 'DEVICE_AGE_TEXT', 'DEVICE_EVALUATED_BY_MANUFACTUR']

df_BI = pd.read_csv('1_BI_KEY_list.txt', header=None, names=['MDR_REPORT_KEY'])

df_list = []
for s in FOI_DEV_LIST:
    df = pd.read_csv('foidev/foidev'+s+'.txt', sep='|', header=0, encoding='ISO-8859-1', error_bad_lines=False)
    df_list.append(df)
df_BASELINE = pd.concat(df_list, axis=0)
del df_list

df_BI_BASELINE = df_BASELINE.merge(df_BI, on=['MDR_REPORT_KEY'], how='inner')

line_num = df_BI_BASELINE.shape[0]
for b in baseline_col:
    m = df_BI_BASELINE[b].isnull().sum()
    print(b, 'missing :', m, 'out of', line_num, ',', m/line_num)

for b in baseline_col:
    del df_BI_BASELINE[b]
df_list = [df_BI_BASELINE]
for s in FOI_DEV_LIST_2:
    df = pd.read_csv('foidev/foidev'+s+'.txt', sep='|', header=0, encoding='ISO-8859-1', error_bad_lines=False)
    df_list.append(df)
df = pd.concat(df_list, axis=0)
del df_list
df = df.merge(df_BI, on=['MDR_REPORT_KEY'], how='inner')

line_num = df.shape[0]
for b in main_col:
    n = df[b].isnull().sum()
    print(b, 'missing :', n, 'out of', line_num, ',', n/line_num)

u = list(df.BRAND_NAME.unique())
print('Distinct brand name:', len(u), 'out of', df['BRAND_NAME'].notnull().sum())
u = list(df.MODEL_NUMBER.unique())
print('Distinct model #:', len(u), 'out of', df['MODEL_NUMBER'].notnull().sum())

df['DATE_RECEIVED'] = pd.to_datetime(df['DATE_RECEIVED'])

def slice(d):
    if (d < datetime(year=1994, month=1, day=1)) | (d >= datetime(year=2017, month=2, day=1)):
        return True
    else:
        return False

df['filter'] = df['DATE_RECEIVED'].map(slice)
wrong_list = list(df.loc[df['filter']==True, :].MDR_REPORT_KEY.unique())
len(wrong_list)
