==> 11_Word_Frequency.ipynb <==
import os
from string import punctuation
import pandas as pd
import nltk
from nltk.tokenize import RegexpTokenizer
from nltk.stem.wordnet import WordNetLemmatizer

lmtzr = WordNetLemmatizer()
tokenizer = RegexpTokenizer(r'\w+')
stopwords = nltk.corpus.stopwords.words('english')

def processLine(line):
    tokens = tokenizer.tokenize(line)
    output = []
    for t in tokens:
        t = lmtzr.lemmatize(t)
        if (len(t) <= 2) or (len(t) >= 15) or (t in stopwords) or (t.isdigit()):
            continue
        output.append(t)
    return ' '.join(output)

manufacture = ['MENTOR', 'ALLERGAN', 'IDEAL', 'SIENTRA', 'MCGHAN', 'INAMED', 'SILIMED', 'NAGOR']
fill_type = ['SALINE', 'SILICONE', 'GEL', 'COHESIVE']
surface_type = ['SMOOTH', 'TEXTURED', 'BIOCELL']
implantation_indication = ['AUGMENTATION', 'RECONSTRUCTION', 'COSMETIC']
ALCL = ['ALCL', 'ANAPLASTIC LARGE CELL LYMPHOMA', 'LYMPHOMA', 'T-CELL LYMPHOMA', 'B-CELL LYMPHOMA']
side = ['LEFT', 'RIGHT']
symptom = ['breast pain', 'breast swell', 'breast cyst', 'breast calcif', 'capsular contracture',
           'lymph node enlarge', 'firmness of breast', 'hematoma', 'mass', 'lump', 'rupture', 'deflat',
           'infect', 'abscess', 'leukopenia', 'nodule', 'skin discolor', 'skin lesion', 'seroma',
           'effusion', 'fluid']
word_list = [manufacture, fill_type, surface_type, implantation_indication, ALCL, side, symptom]

MDR_KEY_LIST = []
with open('Data/No time filter/FWM_Key_no_time_filter.txt') as fh:
    lines = fh.readlines()
    for line in lines:
        MDR_KEY_LIST.append(line.rstrip('\n'))
with open('Data/No time filter/FTR_Key_no_time_filter.txt') as fh:
    lines = fh.readlines()
    for line in lines:
        MDR_KEY_LIST.append(line.rstrip('\n'))

processed = []
for l in word_list:
    for w in l:
        p = processLine(w)
        print(w, ':', p)

def filtering(s):
    if (s == 'FWM') | (s == 'FTR'):
        return True
    else:
        return False

FOI_DEV_LIST = ['', 'Add', 'Change', 'thru1997']
for i in range(1998, 2017):
    FOI_DEV_LIST.append(str(i))

df_list = []
for s in FOI_DEV_LIST:
    df = pd.read_csv('foidev/foidev'+s+'.txt', sep='|', header=0, encoding='ISO-8859-1', error_bad_lines=False)
    df_list.append(df)
df_BASELINE = pd.concat(df_list, axis=0)
del df_list

df_BASELINE['filter'] = df_BASELINE['DEVICE_REPORT_PRODUCT_CODE'].map(filtering)
df_BI_DEV = df_BASELINE.loc[df_BASELINE['filter']==True, :]

print(df_BI_DEV.shape[0])
# del df_BI_DEV['filter']
for n in df_BI_DEV.columns.values:
    if 'BASELINE' in n:
        del df_BI_DEV[n]
df_BI_DEV.to_csv('WF/DEV_BI_FULL_TABLE.txt', header=True, index=False, sep='|')

print('Manufacturer word frequecy:\n')
for m in manufacture:
    counter = 0
    fh = open('WF/manufacturer/'+m+'.txt', 'w')
    for index, value in df_BI_DEV.iterrows():
        l = str(value['MANUFACTURER_D_NAME']).split(' ')
        if m in l:
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

FOI_TEXT_LIST = ['', 'Add', 'Change', 'thru1995']
for i in range(1996, 2017):
    FOI_TEXT_LIST.append(str(i))

df_BI = pd.DataFrame(df_BI_DEV['MDR_REPORT_KEY'])
df_BI_TEXT = pd.DataFrame()
for s in FOI_TEXT_LIST:
    df = pd.read_csv('foitext/foitext'+s+'.txt', sep='|', header=0, encoding='ISO-8859-1', error_bad_lines=False)
    temp = df.merge(df_BI, on=['MDR_REPORT_KEY'], how='inner')
    df_BI_TEXT = pd.concat([df_BI_TEXT, temp], axis=0)
    del temp

print(df_BI_TEXT.shape[0])
df_BI_TEXT.head()

df_BI_TEXT = df_BI_TEXT.drop_duplicates('MDR_TEXT_KEY')
print(df_BI_TEXT.shape[0])
df_BI_TEXT.to_csv('WF/TEXT_FULL_BI_LIST.txt', header=True, index=False, sep='|')
df_BI_TEXT.head()

punctuations = '''!()[]{};:'"\,<>./?@#$%^&*_~'''

def processing_text(s):
    s = str(s)
    no_punct = ""
    for char in s:
        if char not in punctuations:
            no_punct = no_punct + char
    l = no_punct.split(' ')
    output = []
    for w in l:
        if (w in stopwords) or (w.isdigit()):
            continue
        output.append(w)
    return ' '.join(output)

df_BI_TEXT['FOI_TEXT'] = df_BI_TEXT['FOI_TEXT'].map(processing_text)
df_BI_TEXT['FOI_TEXT'].head()

print('Fill type word frequecy:\n')
for m in fill_type:
    counter = 0
    fh = open('WF/fill_type/'+m+'.txt', 'w')
    for index, value in df_BI_TEXT.iterrows():
        l = str(value['FOI_TEXT']).split(' ')
        if m in l:
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

print('Surface type word frequecy:\n')
for m in surface_type:
    counter = 0
    fh = open('WF/surface_type/'+m+'.txt', 'w')
    for index, value in df_BI_TEXT.iterrows():
        l = str(value['FOI_TEXT']).split(' ')
        if m in l:
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

print('Implantation indication word frequecy:\n')
for m in implantation_indication:
    counter = 0
    fh = open('WF/implantation_indication/'+m+'.txt', 'w')
    for index, value in df_BI_TEXT.iterrows():
        l = str(value['FOI_TEXT']).split(' ')
        if m in l:
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

print('ALCL word frequecy:\n')
for m in ALCL:
    counter = 0
    fh = open('WF/ALCL/'+m+'.txt', 'w')
    for index, value in df_BI_TEXT.iterrows():
        if m in str(value['FOI_TEXT']):
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

print('Side word frequecy:\n')
for m in side:
    counter = 0
    fh = open('WF/side/'+m+'.txt', 'w')
    for index, value in df_BI_TEXT.iterrows():
        if m in str(value['FOI_TEXT']):
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

print('Symptom word frequecy:\n')
for m in symptom:
    m = m.upper()
    counter = 0
    fh = open('WF/symptom/'+m+'.txt', 'w')
    for index, value in df_BI_TEXT.iterrows():
        if m in str(value['FOI_TEXT']):
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

df_BI_DEV['BRAND_NAME'] = df_BI_DEV['BRAND_NAME'].map(processing_text)
df_BI_DEV['BRAND_NAME'].head()

print('Surface type word frequecy (brand name):\n')
for m in surface_type:
    counter = 0
    fh = open('WF/surface_type_BRAND_NAME/'+m+'.txt', 'w')
    for index, value in df_BI_DEV.iterrows():
        if m in str(value['BRAND_NAME']):
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

print('Fill type word frequecy (brand name):\n')
for m in fill_type:
    counter = 0
    fh = open('WF/fill_type_BRAND_NAME/'+m+'.txt', 'w')
    for index, value in df_BI_DEV.iterrows():
        if m in str(value['BRAND_NAME']):
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

print('Surface type word frequecy (generic name):\n')
for m in surface_type:
    counter = 0
    fh = open('WF/surface_type_GENERIC_NAME/'+m+'.txt', 'w')
    for index, value in df_BI_DEV.iterrows():
        if m in str(value['GENERIC_NAME']):
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()

print('Fill type word frequecy (generic name):\n')
for m in fill_type:
    counter = 0
    fh = open('WF/fill_type_GENERIC_NAME/'+m+'.txt', 'w')
    for index, value in df_BI_DEV.iterrows():
        if m in str(value['GENERIC_NAME']):
            fh.write(str(value['MDR_REPORT_KEY']) + '\n')
            counter += 1
    print(m, ':', counter)
    fh.close()
