==> 13_Word_Frequency_&_Vector.ipynb <==
import os
from string import punctuation
import pandas as pd
import numpy as np
import nltk
from nltk.tokenize import RegexpTokenizer
from nltk.stem.wordnet import WordNetLemmatizer

fill_type = ['SALINE', 'SILICONE', 'GEL', 'COHESIVE']
surface_type = ['SMOOTH', 'TEXTURED', 'BIOCELL', 'MICROCELL', 'POLYURETHANE']
implantation_indication = ['AUGMENTATION', 'RECONSTRUCTION', 'COSMETIC', 'REVISION', 'AUGMENTATION REVISION',
                          'RECONSTRUCTION REVISION']
ALCL = ['ALCL', 'ANAPLASTIC LARGE CELL LYMPHOMA', 'LYMPHOMA', 'T-CELL LYMPHOMA', 'B-CELL LYMPHOMA',
        'CANCER', 'TUMOR', 'SUSPECT', 'CONFIRMED', 'SUSPECT ALCL', 'CONFIRMED ALCL']
side = ['LEFT', 'RIGHT', 'BILATERAL', 'BOTH SIDES', 'LEFT+RIGHT']
biomarker = ['CD 30', 'CD30', 'ALK', 'NEGATIVE', 'POSITIVE', 'CD30-', 'CD30+', 'ALK-', 'ALK+', 'CD30 NEGATIVE',
            'CD30 POSITIVE', 'ALK NEGATIVE', 'ALK POSITIVE']
symptom = ['breast pain', 'breast swell', 'breast cyst', 'breast calcif', 'capsular contracture',
           'lymph node enlarge', 'firmness of breast', 'hematoma', 'mass', 'lump', 'rupture', 'deflat',
           'infect', 'abscess', 'leukopenia', 'nodule', 'skin discolor', 'skin lesion', 'seroma',
           'effusion', 'fluid', 'edema', 'leak', 'red', 'tender', 'erythema', 'asymmetr', 'breast enlarge']

df_temp = pd.read_excel('2_foidev_column_full_list_standardized.xlsx', sheetname='foidev_MANUFACTURER_full_list')

punctuations = '''!()[]{};:'"\,<>./?@#$%^&*_~'''
stopwords = nltk.corpus.stopwords.words('english')

def processing_text(s):
    s = str(s)
    s = s.lower()
    no_punct = ""
    for char in s:
        if char not in punctuations:
            no_punct = no_punct + char
    if ('alcl-' in no_punct) | ('gel-' in no_punct):
        no_punct = no_punct.replace('alcl-', 'alcl ')
        no_punct = no_punct.replace('alcl-', 'gel ')
    l = no_punct.split(' ')
    output = []
    for w in l:
        if w in stopwords:
            continue
        output.append(w)
    return ' '.join(output).upper()

def convert_manufacturer(m):
    std_manufacturer = df_temp.loc[df_temp['Value']==m, 'Standardized term']
    if len(std_manufacturer) == 0:
        return m
    else:
        std_manufacturer = str(std_manufacturer.values[0])
        std_manufacturer = std_manufacturer.upper()
        return std_manufacturer

df_BI_DEV = pd.read_csv('WF/DEV_BI_FULL_TABLE.txt', sep='|', header=0, encoding='ISO-8859-1', error_bad_lines=False)
df_BI_DEV = df_BI_DEV.drop_duplicates()
df_BI_TEXT = pd.read_csv('WF/TEXT_FULL_BI_LIST.txt', sep='|', header=0, encoding='ISO-8859-1', error_bad_lines=False)
df_BI_TEXT = df_BI_TEXT.drop_duplicates('MDR_TEXT_KEY')
print('Device #:', df_BI_DEV.shape[0])
print('Text #:', df_BI_TEXT.shape[0])
df_BI = df_BI_TEXT.merge(df_BI_DEV, on='MDR_REPORT_KEY', how='inner')
df_BI['MANUFACTURER_D_NAME'] = df_BI['MANUFACTURER_D_NAME'].astype(str)
df_BI['MANUFACTURER_D_NAME'] = df_BI['MANUFACTURER_D_NAME'].map(convert_manufacturer)
df_BI['BRAND_NAME'] = df_BI['BRAND_NAME'].astype(str)
df_BI['GENERIC_NAME'] = df_BI['GENERIC_NAME'].astype(str)
df_BI['FOI_TEXT'] = df_BI['FOI_TEXT'].astype(str)
df_BI = df_BI.drop_duplicates()
df_BI['FOI_TEXT'] = df_BI['FOI_TEXT'].map(processing_text)
print('Full table #:', df_BI.shape[0])

def text_search(text, k):
    l = text.split(' ')
    if k == 'B-CELL LYMPHOMA':
        if (k in text) | ('B CELL LYMPHOMA' in text):
            return True
        else:
            return False
    elif ' ' in k:
        if k in text:
            return True
        else:
            return False
    elif k == 'LEFT+RIGHT':
        if ('LEFT' in text) & ('RIGHT' in text):
            return True
        else:
            return False
    else:
        for i in l:
            if k in i:
                return True
        return False

def filter_merge(f1, f2, f3):
    return (f1 | f2 | f3)

df_vector = pd.DataFrame()
df_vector['MDR_REPORT_KEY'] = df_BI['MDR_REPORT_KEY']
df_vector['MANUFACTURER_D_NAME'] = df_BI['MANUFACTURER_D_NAME']
df_vector['BRAND_NAME'] = df_BI['BRAND_NAME']
df_vector['GENERIC_NAME'] = df_BI['GENERIC_NAME']
df_vector['FOI_TEXT'] = df_BI['FOI_TEXT']

df_temp = pd.read_excel('2_foidev_column_full_list_standardized.xlsx', sheetname='foidev_MANUFACTURER_full_list')

print('Manufacturer word frequecy:\n')
manu_list = list(set(list(df_BI['MANUFACTURER_D_NAME'])))
manu_list.remove('nan')
manu_list.remove('NAN')
manu_list.remove('UNK')
for m in manu_list:
    df_BI['filter'] = np.vectorize(text_search)(df_BI['MANUFACTURER_D_NAME'], m)
    temp = df_BI.loc[df_BI['filter']==True, :]
    df_vector[m] = np.nan
    df_vector.loc[df_BI['filter']==True, m] = 1
    try:
        df_vector[m] = df_vector[m].fillna(0)
        df_vector[m] = df_vector[m].astype(int)
    except:
        continue
    del df_BI['filter']
    temp.to_csv('WF_FULL_TABLE/manufacturer/'+m+'.txt', header=True, sep='|', index=False)
    print(m, ':', temp.shape[0])
    del temp

print('Surface type word frequecy:\n')
for s in surface_type:
    df_BI['filter_1'] = np.vectorize(text_search)(df_BI['BRAND_NAME'], s)
    df_BI['filter_2'] = np.vectorize(text_search)(df_BI['GENERIC_NAME'], s)
    df_BI['filter_3'] = np.vectorize(text_search)(df_BI['FOI_TEXT'], s)
    df_BI['filter'] = np.vectorize(filter_merge)(df_BI['filter_1'], df_BI['filter_2'], df_BI['filter_3'])
    temp = df_BI.loc[df_BI['filter']==True, :]
    df_vector[s] = np.nan
    df_vector.loc[df_BI['filter']==True, s] = 1
    try:
        df_vector[s] = df_vector[s].fillna(0)
        df_vector[s] = df_vector[s].astype(int)
    except:
        continue
    del df_BI['filter'], df_BI['filter_1'], df_BI['filter_2'], df_BI['filter_3']
    temp.to_csv('WF_FULL_TABLE/surface/'+s+'.txt', header=True, sep='|', index=False)
    print(s, ':', temp.shape[0])
    del temp

print('Fill type word frequecy:\n')
for f in fill_type:
    df_BI['filter_1'] = np.vectorize(text_search)(df_BI['BRAND_NAME'], f)
    df_BI['filter_2'] = np.vectorize(text_search)(df_BI['GENERIC_NAME'], f)
    df_BI['filter_3'] = np.vectorize(text_search)(df_BI['FOI_TEXT'], f)
    df_BI['filter'] = np.vectorize(filter_merge)(df_BI['filter_1'], df_BI['filter_2'], df_BI['filter_3'])
    temp = df_BI.loc[df_BI['filter']==True, :]
    df_vector[f] = np.nan
    df_vector.loc[df_BI['filter']==True, f] = 1
    try:
        df_vector[f] = df_vector[f].fillna(0)
        df_vector[f] = df_vector[f].astype(int)
    except:
        continue
    del df_BI['filter'], df_BI['filter_1'], df_BI['filter_2'], df_BI['filter_3']
    temp.to_csv('WF_FULL_TABLE/fill/'+f+'.txt', header=True, sep='|', index=False)
    print(f, ':', temp.shape[0])
    del temp

print('Implantation indication word frequecy:\n')
for i in implantation_indication:
    df_BI['filter'] = np.vectorize(text_search)(df_BI['FOI_TEXT'], i)
    temp = df_BI.loc[df_BI['filter']==True, :]
    df_vector[i] = np.nan
    df_vector.loc[df_BI['filter']==True, i] = 1
    try:
        df_vector[i] = df_vector[i].fillna(0)
        df_vector[i] = df_vector[i].astype(int)
    except:
        continue
    del df_BI['filter']
    temp.to_csv('WF_FULL_TABLE/implantation_indication/'+i+'.txt', header=True, sep='|', index=False)
    print(i, ':', temp.shape[0])
    del temp

print('ALCL word frequecy:\n')
for m in ALCL:
    df_BI['filter'] = np.vectorize(text_search)(df_BI['FOI_TEXT'], m)
    temp = df_BI.loc[df_BI['filter']==True, :]
    df_vector[m] = np.nan
    df_vector.loc[df_BI['filter']==True, m] = 1
    try:
        df_vector[m] = df_vector[m].fillna(0)
        df_vector[m] = df_vector[m].astype(int)
    except:
        continue
    del df_BI['filter']
    temp.to_csv('WF_FULL_TABLE/ALCL/'+m+'.txt', header=True, sep='|', index=False)
    print(m, ':', temp.shape[0])
    del temp

print('Side word frequecy:\n')
for s in side:
    df_BI['filter'] = np.vectorize(text_search)(df_BI['FOI_TEXT'], s)
    temp = df_BI.loc[df_BI['filter']==True, :]
    df_vector[s] = np.nan
    df_vector.loc[df_BI['filter']==True, s] = 1
    try:
        df_vector[s] = df_vector[s].fillna(0)
        df_vector[s] = df_vector[s].astype(int)
    except:
        continue
    del df_BI['filter']
    temp.to_csv('WF_FULL_TABLE/side/'+s+'.txt', header=True, sep='|', index=False)
    print(s, ':', temp.shape[0])
    del temp

print('Biomarker word frequecy:\n')
for b in biomarker:
    df_BI['filter'] = np.vectorize(text_search)(df_BI['FOI_TEXT'], b)
    temp = df_BI.loc[df_BI['filter']==True, :]
    df_vector[b] = np.nan
    df_vector.loc[df_BI['filter']==True, b] = 1
    try:
        df_vector[b] = df_vector[b].fillna(0)
        df_vector[b] = df_vector[b].astype(int)
    except:
        continue
    del df_BI['filter']
    temp.to_csv('WF_FULL_TABLE/biomarker/'+b+'.txt', header=True, sep='|', index=False)
    print(b, ':', temp.shape[0])
    del temp

print('Symptom word frequecy:\n')
for s in symptom:
    s = s.upper()
    df_BI['filter'] = np.vectorize(text_search)(df_BI['FOI_TEXT'], s)
    temp = df_BI.loc[df_BI['filter']==True, :]
    df_vector[s] = np.nan
    df_vector.loc[df_BI['filter']==True, s] = 1
    try:
        df_vector[s] = df_vector[s].fillna(0)
        df_vector[s] = df_vector[s].astype(int)
    except:
        continue
    del df_BI['filter']
    temp.to_csv('WF_FULL_TABLE/symptom/'+s+'.txt', header=True, sep='|', index=False)
    print(s, ':', temp.shape[0])
    del temp

df_vector.to_csv('keyword_vector.txt', sep='|', header=True, index=False)
df_vector.head(20)

print('Is there saline and silicone at the same time?')
fill_conflict = df_vector.loc[(df_vector['SALINE']==1) & (df_vector['SILICONE']==1), :]
fill_conflict.to_excel('fill_conflict.xlsx', 'fill_conflict', index=False)
fill_conflict.head()

print('Is there smooth and textured at the same time?')
surface_conflict = df_vector.loc[(df_vector['SMOOTH']==1) & (df_vector['TEXTURED']==1), :]
surface_conflict.to_excel('surface_conflict.xlsx', 'surface_conflict', index=False)
surface_conflict.head()

print('Is there left and right at the same time?')
side_conflict = df_vector.loc[df_vector['LEFT+RIGHT']==1, :]
side_conflict.to_excel('side_conflict.xlsx', 'side_conflict', index=False)
side_conflict.head()
